Move capture/ to runner/
