--- src/runtime_src/core/common/capture/CMakeLists.txt
@@ -1,12 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
-add_library(core_common_capture_objects OBJECT
-  frames.cpp
-  )
-
-target_include_directories(core_common_capture_objects
-  PRIVATE
-  ${XRT_SOURCE_DIR}/runtime_src
-  ${XRT_SOURCE_DIR}/runtime_src/core/common/elf
-  )
-
