+++ src/runtime_src/core/common/runner/CMakeLists.txt
@@ -1,6 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright (C) 2024 Advanced Micro Devices, Inc. All rights reserved.
 add_library(core_common_runner_objects OBJECT
+  capture.cpp
   repo.cpp
   runner.cpp
   cpu.cpp
@@ -9,6 +10,7 @@ add_library(core_common_runner_objects OBJECT
 target_include_directories(core_common_runner_objects
   PRIVATE
   ${XRT_SOURCE_DIR}/runtime_src
+  ${XRT_SOURCE_DIR}/runtime_src/core/common/elf
   )
 
 xrt_configure_version_file(xrt-runner APP)
