+++ src/runtime_src/core/common/runner/capture.h
@@ -1,11 +1,10 @@
-
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
-#ifndef XRT_COMMON_CAPTURE_CAPTURE_H_
-#define XRT_COMMON_CAPTURE_CAPTURE_H_
+#ifndef XRT_COMMON_RUNNER_CAPTURE_H_
+#define XRT_COMMON_RUNNER_CAPTURE_H_
 #include "xrt/detail/config.h"
 #include "core/common/config_reader.h"
-#include "core/common/capture/fn_fwd.h"
+#include "core/common/runner/detail/capture_fnfwd.h"
 
 namespace xrt_core::capture {
 
@@ -25,7 +24,7 @@ is_enabled()
 // No-op when disabled, actual capture when enabled
 #define XRT_REPLAY_CAPTURE(fn, ...) \
   do { if (XRT_UNLIKELY(xrt_core::capture::is_enabled())) { \
-      xrt_core::capture::fn(__VA_ARGS__);           \
+      xrt_core::capture::detail::fn(__VA_ARGS__);           \
     } } while(0)
 
 } // namespace xrt_core::capture
